+++ AutomatedLabCore/functions/Azure/Get-LabAzureDefaultResourceGroup.ps1
@@ -7,6 +7,8 @@
 
     Update-LabAzureSettings
 
+    if ($script:lab.AzureSettings.DefaultResourceGroup) { return $script:lab.AzureSettings.DefaultResourceGroup }
+
     $script:lab.AzureSettings.ResourceGroups | Where-Object ResourceGroupName -eq $script:lab.Name
 
     Write-LogFunctionExit