+++ AutomatedLabWorker/functions/AzureWorkerVirtualMachines/New-LabAzureResourceGroupDeployment.ps1
@@ -367,7 +367,6 @@
                 )
                 type       = "Microsoft.Network/virtualNetworks/virtualNetworkPeerings"
                 name       = "$($network.ResourceName)/$($network.ResourceName)To$($peer)"
-                location   = "[resourceGroup().location]"
                 properties = @{
                     allowVirtualNetworkAccess = $true
                     allowForwardedTraffic     = $false
@@ -375,6 +374,46 @@
                     useRemoteGateways         = $false
                     remoteVirtualNetwork      = @{
                         id = "[resourceId('Microsoft.Network/virtualNetworks', '$peer')]"
+                    }
+                }
+            }
+            $template.Resources += @{
+                apiVersion = $apiVersions['VirtualNetworkApi']
+                dependsOn  = @(
+                    "[resourceId('Microsoft.Network/virtualNetworks', '$($network.ResourceName)')]"
+                    "[resourceId('Microsoft.Network/virtualNetworks', '$($peer)')]"
+                )
+                type       = "Microsoft.Network/virtualNetworks/virtualNetworkPeerings"
+                name       = "$($peer)/$($peer)To$($network.ResourceName)"
+                properties = @{
+                    allowVirtualNetworkAccess = $true
+                    allowForwardedTraffic     = $false
+                    allowGatewayTransit       = $false
+                    useRemoteGateways         = $false
+                    remoteVirtualNetwork      = @{
+                        id = "[resourceId('Microsoft.Network/virtualNetworks', '$($network.ResourceName)')]"
+                    }
+                }
+            }
+        }
+
+        foreach ($externalPeer in $network.PeeringVnetResourceIds) {
+            $peerName = $externalPeer -split '/' | Select-Object -Last 1
+            Write-ScreenInfo -Type Verbose -Message ('Adding peering from {0} to {1} to VNet template' -f $network.ResourceName, $peerName)
+            $template.Resources += @{
+                apiVersion = $apiVersions['VirtualNetworkApi']
+                dependsOn  = @(
+                    "[resourceId('Microsoft.Network/virtualNetworks', '$($network.ResourceName)')]"
+                )
+                type       = "Microsoft.Network/virtualNetworks/virtualNetworkPeerings"
+                name       = "$($network.ResourceName)/$($network.ResourceName)To$($peerName)"
+                properties = @{
+                    allowVirtualNetworkAccess = $true
+                    allowForwardedTraffic     = $false
+                    allowGatewayTransit       = $false
+                    useRemoteGateways         = $false
+                    remoteVirtualNetwork      = @{
+                        id = $externalPeer
                     }
                 }
             }
@@ -670,23 +709,6 @@
         $vmSize = Get-LWAzureVmSize -Machine $Machine
         $imageRef = Get-LWAzureSku -Machine $machine
 
-        if (($Machine.VmGeneration -eq 2 -and $vmSize.Gen2Supported) -or ($vmSize.Gen2Supported -and -not $vmSize.Gen1Supported))
-        {
-            $pattern = '{0}(-g2$|gen2|-gensecond$)' -f $imageRef.sku # Yes, why should the image names be consistent? Also of course we don't need a damn VMGeneration property...
-            $newImage = $lab.AzureSettings.VMImages | Where-Object { $_.PublisherName -eq $imageref.Publisher -and $_.Offer -eq $imageref.Offer -and $_.Skus -match $pattern }
-            if (-not $newImage)
-            {
-                throw "Selected VM size $vmSize for $Machine only suppports G2 VMs, however no matching Generation 2 image was found for your selection: Publisher $($imageRef.publisher), offer $($imageRef.offer), sku $($imageRef.sku)!"
-            }
-
-            $imageRef = @{
-                publisher = $newImage.PublisherName
-                version   = $newImage.Version
-                offer     = $newImage.Offer
-                sku       = $newImage.Skus
-            }
-        }
-
         if (-not $vmSize)
         {
             throw "No valid VM size found for '$Machine'. For a list of available role sizes, use the command 'Get-LabAzureAvailableRoleSize -LocationName $($lab.AzureSettings.DefaultLocation.Location)'"
@@ -694,7 +716,6 @@
 
         Write-ScreenInfo -Type Verbose -Message "Adding $Machine with size $vmSize, publisher $($imageRef.publisher), offer $($imageRef.offer), sku $($imageRef.sku)!"
 
-        $machNet = Get-LabVirtualNetworkDefinition -Name $machine.Network[0]
         $machTemplate = @{
             name       = $machine.ResourceName
             tags       = @{